+++ R/0_imports.R
@@ -30,6 +30,8 @@ utils::globalVariables(
     "estimate",                                                  # lowerimpute.R
     ".row",                                                          # integer.R
     "denom",                                                           # ratio.R
-    "component", "denom", "id", "value"                                  # ica.R
+    "component", "denom", "id", "value",                                 # ica.R
+    "training", "x_names", "y_names",                                    # pls.R
+    ".order_1", ".order_2"                                           # recipes.R
   )
 )