+++ R/downsample.R
@@ -1,8 +1,12 @@
 #' Down-Sample a Data Set Based on a Factor Variable
 #'
-#' `step_downsample` creates a *specification* of a recipe
-#'  step that will remove rows of a data set to make the occurrence
-#'  of levels in a specific factor level equal.
+#' @description
+#' \if{html}{\figure{lifecycle-soft-deprecated.svg}{alt="lifecycle-soft-deprecated"}}
+#'
+#' `step_downsample` is now available as `themis::step_downsample()`. This
+#'  function creates a *specification* of a recipe step that will remove
+#'  rows of a data set to make the occurrence of levels in a specific factor
+#'  level equal.
 #'
 #' @inheritParams step_center
 #' @param ... One or more selector functions to choose which
@@ -20,7 +24,7 @@
 #'  frequency as the least occurring level. A value of 2 would mean
 #'  that the majority levels will have (at most) (approximately)
 #'  twice as many rows than the minority level.
-#' @param ratio Depracated argument; same as `under_ratio`
+#' @param ratio Deprecated argument; same as `under_ratio`
 #' @param target An integer that will be used to subsample. This
 #'  should not be set by the user and will be populated by `prep`.
 #' @param seed An integer that will be used as the seed when downsampling.
@@ -80,6 +84,10 @@ step_downsample <-
   function(recipe, ...,  under_ratio = 1, ratio = NA, role = NA, trained = FALSE,
            column = NULL, target = NA, skip = TRUE,
            seed = sample.int(10^5, 1), id = rand_id("downsample")) {
+
+    lifecycle::deprecate_soft("0.1.13",
+                              "recipes::step_downsample()",
+                              "themis::step_downsample()")
 
     if (!is.na(ratio) & all(under_ratio != ratio)) {
       message(